% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_statsnz.R
\name{read_statsnz}
\alias{read_statsnz}
\title{Read CSVs from StatsNZ}
\usage{
read_statsnz(name, subname = NULL, path = tempdir())
}
\arguments{
\item{name}{A name of a StatsNZ CSV or ZIP, or a unique fragment thereof}

\item{subname}{Used when \code{name} refers to a ZIP. \code{subname} is the name,
or a unique fragment thereof, of a file within the ZIP.}

\item{path}{Directory in which to save downloaded file(s)}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function downloads and imports CSVs from StatsNZ
}
\examples{
\dontrun{
read_statsnz("national-labour-force-projections")

read_statsnz("labour-market-statistics", subname = "lci")
}
}
