% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_available_csvs.R
\name{show_available_csvs}
\alias{show_available_csvs}
\title{List CSVs available to download from the StatsNZ website}
\usage{
show_available_csvs(refresh = TRUE)
}
\arguments{
\item{refresh}{Logical; \code{TRUE} by default. If \code{TRUE}, a current list of
available CSVs will be scraped from the StatsNZ website. If \code{FALSE}, a
cached internal list of CSVs will be returned.}
}
\value{
a data.frame (tibble) containing three columns: the \code{title},
\code{name}, \code{url}
and \code{ext} (file extension) for a range of CSVs available to download from
StatsNZ.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
show_available_csvs()
}
