% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_statsnz.R
\name{read_statsnz}
\alias{read_statsnz}
\title{Read CSVs from StatsNZ}
\usage{
read_statsnz(name, subname = NULL, path = tempdir(), check_local = TRUE)
}
\arguments{
\item{name}{A name of a StatsNZ CSV or ZIP, or a unique fragment thereof}

\item{subname}{Used when \code{name} refers to a ZIP that contains more than one
CSV. \code{subname} is the name,
or a unique fragment thereof, of a CSV file within the ZIP.}

\item{path}{Directory in which to save downloaded file(s)}

\item{check_local}{Logical; \code{TRUE} by default. If \code{TRUE}, \code{read_statsnz()}
will look in \code{path} for a file with the same filename as the one you've
requested. If it exists, it will be loaded. If not, the file will be
downloaded from StatsNZ. If \code{FALSE}, the file will always be
downloaded from StatsNZ.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function downloads and imports CSVs from StatsNZ
}
\examples{
\dontrun{
read_statsnz("national-labour-force-projections")

read_statsnz("labour-market-statistics", subname = "lci")
}
}
